[Shared Utils] Centralize error utils in SDK and refactor imports
