+++ connectors/src/types/api.ts
@@ -159,20 +159,4 @@ export function isScheduleAlreadyRunning(err: unknown) {
   );
 }
 
-function errorToString(error: unknown): string {
-  if (error instanceof Error) {
-    return error.message;
-  } else if (typeof error === "string") {
-    return error;
-  }
-
-  return JSON.stringify(error);
-}
-
-export function normalizeError(error: unknown): Error {
-  if (error instanceof Error) {
-    return error;
-  }
-
-  return new Error(errorToString(error));
-}
+export { errorToString, normalizeError } from "@dust-tt/client";
