+++ front/types/shared/utils/error_utils.ts
@@ -1,22 +1,5 @@
 import { DustError } from "@app/lib/error";
-
-export function errorToString(error: unknown): string {
-  if (error instanceof Error) {
-    return error.message;
-  } else if (typeof error === "string") {
-    return error;
-  }
-
-  return JSON.stringify(error);
-}
-
-export function normalizeError(error: unknown): Error {
-  if (error instanceof Error) {
-    return error;
-  }
-
-  return new Error(errorToString(error));
-}
+import { errorToString, normalizeError } from "@dust-tt/client";
 
 export function normalizeAsInternalDustError(
   error: unknown
